third_party/blink/web_tests/virtual/select-parser-relaxation-opt-out/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-datalist-options-idl.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Option elements should work if they are a descendant of a selects wrapper div.
  Cannot read properties of null (reading 'appendChild')
[FAIL] Options in wrapper div should still work when the multiple attribute is added.
  assert_equals: select.length expected 6 but got 5
[FAIL] Options in wrapper div in multiple should work after re-parsing and re-attaching.
  assert_equals: select.length expected 6 but got 5
Harness: the test ran to completion.
